refactor: start using methods from framework
- replace methods from render module to methods from framework
- extend all view classes from AbstarctView class from framework
- separate temp constants from work constants
- delete render module from project
