+++ src/constants.js
@@ -2,129 +2,10 @@ const FILM_CARDS_AMOUNT = 12;
 const EXTRA_FILMS_CARDS_AMOUNT = 2;
 const FILMS_PORTION = 5;
 const EMOTIONS = ['smile', 'sleeping', 'puke', 'angry'];
-const TEMP_DESCRIPTION = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.';
-const TEMP_TITLES = [
-  'Naruto: Shippuuden',
-  'One Peace',
-  'Bleach',
-  'Hunter x Hunter',
-  'Claymore',
-  'Berserk',
-  'Evangelion',
-  'Elfen Lied',
-  'Death Note',
-  'Sword Art Online',
-  'Attack on Titan',
-  'One Punch Man',
-  'Mob Psycho 100',
-];
-const TEMP_POSTERS = [
-  'made-for-each-other.png',
-  'popeye-meets-sinbad.png',
-  'sagebrush-trail.jpg',
-  'santa-claus-conquers-the-martians.jpg',
-  'the-dance-of-life.jpg',
-  'the-great-flamarion.jpg',
-  'the-man-with-the-golden-arm.jpg',
-];
-const TEMP_FILM_STUFF = [
-  'Yagami Light',
-  'Naruto Uzumaki',
-  'Gon Fricks',
-  'Monkey D. Luffy',
-  'Kirito',
-  'Eren Jager',
-  'Guts',
-  'Saitama',
-  'Ichigo Kurosaki',
-  'Clare',
-  'Shigeo Kageyama',
-];
-const TEMP_RELEASE_COUNTRIES = [
-  'Russia',
-  'Ukraine',
-  'Japan',
-  'Germany',
-  'USA',
-  'France',
-  'China',
-  'Finland',
-];
-const TEMP_GENRES = [
-  'Action',
-  'Adventure',
-  'Comedy',
-  'Drama',
-  'Fantasy',
-  'Magic',
-  'Supernatural',
-  'Horror',
-  'Mystery',
-  'Psychological',
-  'Romance',
-];
-const TEMP_COMMENTS = [
-  'Ban-Kai!',
-  'Rasengun!',
-  'Rock, Paper, Scissors',
-  'Serious table flip',
-  'Aaaaaaaaaaaa!',
-  'Gomu-Gomu',
-  'Ban-Kai!',
-  'Rasengun!',
-  'Rock, Paper, Scissors',
-  'Serious table flip',
-  'Aaaaaaaaaaaa!',
-  'Gomu-Gomu',
-];
-const TEMP_AGE = {
-  MIN: 0,
-  MAX: 18,
-};
-const TEMP_RATING = {
-  MIN: 0,
-  MAX: 10,
-};
-const TEMP_FILM_STUFF_AMOUNT = {
-  MIN: 1,
-  MAX: 4,
-};
-const TEMP_DATES = {
-  Y_MIN: 1999,
-  Y_MAX: 2022,
-  M_MIN: 1,
-  M_MAX: 12,
-  D_MIN: 1,
-  D_MAX: 30,
-};
-const TEMP_DURATION = {
-  MIN: 0,
-  MAX: 240,
-};
-const TEMP_BOOLEAN = {
-  FALSE: 0,
-  TRUE: 1,
-};
-const TEMP_MAX_COMMENTS = 10;
-
 
 export {
   FILM_CARDS_AMOUNT,
   EXTRA_FILMS_CARDS_AMOUNT,
   FILMS_PORTION,
   EMOTIONS,
-  TEMP_DESCRIPTION,
-  TEMP_TITLES,
-  TEMP_POSTERS,
-  TEMP_FILM_STUFF,
-  TEMP_RELEASE_COUNTRIES,
-  TEMP_GENRES,
-  TEMP_COMMENTS,
-  TEMP_AGE,
-  TEMP_RATING,
-  TEMP_FILM_STUFF_AMOUNT,
-  TEMP_DATES,
-  TEMP_DURATION,
-  TEMP_BOOLEAN,
-  TEMP_MAX_COMMENTS
 };
