+++ src/view/film-card-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createFilmCardElement = (film) => {
   const {title, description, release, poster, genre, runtime, totalRating} = film.filmInfo;
@@ -25,26 +25,15 @@ const createFilmCardElement = (film) => {
         </article>`;
 };
 
-export default class FilmCardView {
-  #element = null;
+export default class FilmCardView extends AbstractView {
   #film = null;
 
   constructor(film) {
+    super();
     this.#film = film;
   }
 
   get template () {
     return createFilmCardElement(this.#film);
-  }
-
-  get element () {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
-  removeElement () {
-    this.#element = null;
   }
 }
