+++ src/view/films-container-view.js
@@ -1,22 +1,9 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const filmsMainContainerElement = () => '<section class="films"></section>';
 
-export default class FilmsContainerView {
-  #element = null;
-
+export default class FilmsContainerView extends AbstractView {
   get template () {
     return filmsMainContainerElement();
-  }
-
-  get element () {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
-  removeElement () {
-    this.#element = null;
   }
 }
