+++ src/view/films-list-container-view.js
@@ -1,22 +1,9 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const filmListContainerElement = () => '<div class="films-list__container"></div>';
 
-export default class FilmsListContainerView {
-  #element = null;
-
+export default class FilmsListContainerView extends AbstractView {
   get template () {
     return filmListContainerElement();
-  }
-
-  get element () {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
-  removeElement () {
-    this.#element = null;
   }
 }
