+++ src/view/films-list-section-view.js
@@ -1,24 +1,11 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createFilmListSectionElement = () => `<section class="films-list">
       <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
     </section>`;
 
-export default class FilmsListSectionView {
-  #element = null;
-
+export default class FilmsListSectionView extends AbstractView {
   get template () {
     return createFilmListSectionElement();
-  }
-
-  get element () {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
-  removeElement () {
-    this.#element = null;
   }
 }
