+++ src/view/no-films-list-section-view.js
@@ -1,28 +1,17 @@
-import {createElement} from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createNoFilmListSectionElement = () => `<section class="films-list">
       <h2 class="films-list__title">There are no movies in our database</h2>`;
 
-export default class NoFilmsListSectionView {
-  #element = null;
+export default class NoFilmsListSectionView extends AbstractView {
   #film = null;
 
   constructor(film) {
+    super();
     this.#film = film;
   }
 
   get template () {
     return createNoFilmListSectionElement(this.#film);
-  }
-
-  get element () {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
-
-  removeElement () {
-    this.#element = null;
   }
 }
